Revert "refactor(websocket): reorganize RedisStore and update module imports …"
This reverts commit 27ad6b77b21f6043be81116885a1e52e857d5874.
