+++ server/websocket/src/bin/websocket.rs
@@ -3,8 +3,8 @@ use std::sync::Arc;
 use tracing::error;
 use uuid::Uuid;
 use websocket::{
-    conf::Config, infrastructure::gcs::GcsStore, infrastructure::redis::RedisStore,
-    pool::BroadcastPool, server::start_server, AppState,
+    conf::Config, infrastructure::gcs::GcsStore, pool::BroadcastPool, server::start_server,
+    storage::redis::RedisStore, AppState,
 };
 
 #[cfg(feature = "auth")]
