+++ server/websocket/src/broadcast/mod.rs
@@ -1,3 +1,4 @@
 pub mod group;
 pub mod pool;
+pub mod sub;
 pub mod types;
