+++ server/websocket/src/domain/entity/mod.rs
@@ -1,1 +1,1 @@
-pub mod doc;
+
