+++ server/websocket/src/domain/mod.rs
@@ -1,3 +1,2 @@
-pub mod entity;
 pub mod repository;
 pub mod value_objects;
