+++ server/websocket/src/domain/repository/mod.rs
@@ -1,2 +1,1 @@
-pub mod doc;
 pub mod kv;
