+++ server/websocket/src/domain/value_objects/http.rs
@@ -4,6 +4,14 @@ use serde::{Deserialize, Serialize};
 use yrs::updates::encoder::Encode;
 
 #[derive(Debug, Clone)]
+pub struct Document {
+    pub id: String,
+    pub updates: Vec<u8>,
+    pub version: u64,
+    pub timestamp: chrono::DateTime<Utc>,
+}
+
+#[derive(Debug, Clone)]
 pub struct HistoryItem {
     pub version: u64,
     pub updates: Vec<u8>,
