+++ server/websocket/src/domain/value_objects/mod.rs
@@ -1,6 +1,4 @@
 pub mod conf;
-pub mod count;
 pub mod http;
 pub mod keys;
 pub mod redis;
-pub mod sub;
