+++ server/websocket/src/infrastructure/mod.rs
@@ -1,3 +1,1 @@
 pub mod gcs;
-pub mod redis;
-pub mod repository;
