--- server/websocket/src/infrastructure/repository/mod.rs
@@ -1,1 +1,0 @@
-// pub mod doc;
