+++ server/websocket/src/storage/mod.rs
@@ -1,0 +1,1 @@
+pub mod redis;
